refactor(ga): separate reused logic and rename js file
